+++ src/primary/FilesystemLibrary.ts
@@ -1,7 +1,29 @@
-// Placeholder for primary library
-// Will be implemented in later prompts
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface PrimaryFilesystemLibrary {
-  // Placeholder
-}
+import { Item } from '@fjell/core';
+import * as Library from '@fjell/lib';
+import { createFilesystemLibrary, FilesystemLibrary } from '../FilesystemLibrary';
+import FSLogger from '../logger';
 
+const logger = FSLogger.get('primary', 'FilesystemLibrary');
+
+/**
+ * Specialized factory for primary items - simpler API
+ */
+export function createPrimaryFilesystemLibrary<
+  V extends Item<S>,
+  S extends string
+>(
+  keyType: S,
+  directory: string,
+  globalDirectory: string,
+  options?: Partial<Library.Options<V, S>>
+): FilesystemLibrary<V, S> {
+  logger.default('createPrimaryFilesystemLibrary', { keyType, directory, globalDirectory });
+
+  return createFilesystemLibrary<V, S>(
+    [keyType],
+    [directory],
+    globalDirectory,
+    options || {},
+    []
+  );
+}